refactor module layout
