+++ types-derive/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "elastic_types_derive"
-version = "0.14.0"
+version = "0.15.0"
 authors = ["Ashley Mannix <ashleymannix@live.com.au>"]
 license = "Apache-2.0"
 description = "Compile-time code generation for Elasticsearch type implementations."
@@ -11,6 +11,6 @@ name = "elastic_types_derive"
 proc-macro = true
 
 [dependencies]
-elastic_types_derive_internals = "~0.2.0"
+elastic_types_derive_internals = { version = "~0.3.0", path = "../types-derive-internals" }
 syn = { version = "~0.11.0", features = ["aster", "visit", "parsing", "full"] }
 quote = "~0.3.0"
