+++ types/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "elastic_types"
-version = "0.14.0"
+version = "0.15.0"
 authors = ["Ashley Mannix <ashleymannix@live.com.au>"]
 license = "Apache-2.0"
 description = "A strongly-typed implementation of Elasticsearch core types and Mapping API."
@@ -15,7 +15,7 @@ geo = "~0.3.0"
 geohash = "~0.3.0"
 geojson = "~0.5.0"
 serde_derive = "~0.9.0"
-elastic_types_derive = "~0.14.0"
+elastic_types_derive = { version = "~0.15.0", path = "../types-derive" }
 
 [dev-dependencies]
 json_str = "^0.*"
