+++ types/src/boolean/impls.rs
@@ -1,10 +1,9 @@
 use std::marker::PhantomData;
 use serde::{Serialize, Deserialize, Serializer, Deserializer};
 use serde::de::{Visitor, Error};
-use super::mapping::{BooleanMapping, DefaultBooleanMapping, BooleanFormat};
-use ::field::FieldType;
+use super::mapping::{BooleanFieldType, BooleanMapping, DefaultBooleanMapping};
 
-impl FieldType<DefaultBooleanMapping, BooleanFormat> for bool {}
+impl BooleanFieldType<DefaultBooleanMapping> for bool {}
 
 /// An Elasticsearch `boolean` with a mapping.
 ///
@@ -74,7 +73,7 @@ impl<M> Boolean<M>
     }
 }
 
-impl<M> FieldType<M, BooleanFormat> for Boolean<M> where M: BooleanMapping {}
+impl<M> BooleanFieldType<M> for Boolean<M> where M: BooleanMapping {}
 
 impl_mapping_type!(bool, Boolean, BooleanMapping);
