+++ types/src/boolean/mod.rs
@@ -33,17 +33,16 @@
 //!
 //! - [Elasticsearch Doc](https://www.elastic.co/guide/en/elasticsearch/reference/current/boolean.html)
 
-#[macro_use]
 pub mod mapping;
 
-mod boolean;
-pub use self::boolean::*;
+mod impls;
+pub use self::impls::*;
 
 pub mod prelude {
     //! Includes all types for the `boolean` type.
     //!
     //! This is a convenience module to make it easy to build mappings for multiple types without too many `use` statements.
 
-    pub use super::boolean::*;
+    pub use super::impls::*;
     pub use super::mapping::*;
 }
