+++ types/src/geo/shape/impls.rs
@@ -1,8 +1,7 @@
 use std::marker::PhantomData;
 use serde::{Serialize, Deserialize, Serializer, Deserializer};
 use geojson::Geometry;
-use super::mapping::*;
-use ::field::FieldType;
+use super::mapping::{GeoShapeFieldType, GeoShapeMapping, DefaultGeoShapeMapping};
 
 /// Geo shape type with a given mapping.
 ///
@@ -68,7 +67,7 @@ impl<M> GeoShape<M>
     }
 }
 
-impl<M> FieldType<M, GeoShapeFormat> for GeoShape<M> where M: GeoShapeMapping {}
+impl<M> GeoShapeFieldType<M> for GeoShape<M> where M: GeoShapeMapping {}
 
 impl_mapping_type!(Geometry, GeoShape, GeoShapeMapping);
