+++ types/src/ip/mod.rs
@@ -34,16 +34,16 @@
 //!
 //! - [Elasticsearch Doc](https://www.elastic.co/guide/en/elasticsearch/reference/current/ip.html)
 
-mod ip;
-
 pub mod mapping;
-pub use self::ip::*;
+
+mod impls;
+pub use self::impls::*;
 
 pub mod prelude {
     //! Includes all types for the `ip` type.
     //!
     //! This is a convenience module to make it easy to build mappings for multiple types without too many `use` statements.
 
-    pub use super::ip::*;
+    pub use super::impls::*;
     pub use super::mapping::*;
 }
