+++ types/src/number/mod.rs
@@ -47,17 +47,16 @@
 //!
 //! - [Elasticsearch Doc](https://www.elastic.co/guide/en/elasticsearch/reference/current/number.html)
 
-#[macro_use]
 pub mod mapping;
 
-mod number;
-pub use self::number::*;
+mod impls;
+pub use self::impls::*;
 
 pub mod prelude {
     //! Includes all types for the `number` type.
     //!
     //! This is a convenience module to make it easy to build mappings for multiple types without too many `use` statements.
 
-    pub use super::number::*;
+    pub use super::impls::*;
     pub use super::mapping::*;
 }
