+++ types/src/string/keyword/mod.rs
@@ -6,15 +6,15 @@
 
 #[macro_use]
 pub mod mapping;
-mod keyword;
 
-pub use self::keyword::*;
+mod impls;
+pub use self::impls::*;
 
 pub mod prelude {
     //! Includes all types for the `keyword` type.
     //!
     //! This is a convenience module to make it easy to build mappings for multiple types without too many `use` statements.
 
-    pub use super::keyword::*;
+    pub use super::impls::*;
     pub use super::mapping::*;
 }
